main

+++ ft_strncmp.c
@@ -14,12 +14,12 @@
 
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+    size_t	i;
 
     i = 0;
     while (s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0' && i < n - 1)
         i++;
     if (n == 0)
         return (0);
-    return (s1[i] - s2[i]);
+    return ((int)(s1[i] - s2[i]));
 }
